Move fpx-directory sub command to the system command
